pyproject.toml
[project]
name = "visual-understanding"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bert-score>=0.3.13",
    "fuzzywuzzy>=0.18.0",
    "gradio-client>=1.10.0",
    "kaleido==0.2.0",
    "nltk>=3.9.1",
    "plotly>=6.0.1",
    "polars>=1.28.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
